refactor: helpers.js deleted, created helpers folder and QRcodeScanner component
